requirements.txt
pip==23.2.1
matplotlib==3.5.3
numpy==1.21.6
pandas
Pillow
plotly
prophet
seaborn
streamlit
yfinance
statsmodel

